chore(ui): Remove Rhai Expression support and related code
