+++ ui/src/features/Editor/components/ParamsDialog/components/ValueEditorDialog/index.tsx
@@ -34,14 +34,9 @@ import { Asset } from "@flow/types";
 
 import { FieldContext } from "../../utils/fieldUtils";
 
-import RhaiCodeEditor, {
-  type RhaiCodeEditorRef,
-} from "./components/RhaiCodeEditor";
-import {
-  TemplateLibraryDialog,
-  TemplatePlaceholderDialog,
-  type ExpressionTemplate,
-} from "./templates";
+import FlowExprCodeEditor, {
+  type FlowExprCodeEditorRef,
+} from "./components/FlowExprCodeEditor";
 
 type Props = {
   open: boolean;
@@ -50,7 +45,7 @@ type Props = {
   onValueSubmit?: (value: any) => void;
 };
 
-export type DialogOptions = "assets" | "cms" | "templates" | undefined;
+export type DialogOptions = "assets" | "cms" | undefined;
 
 const ValueEditorDialog: React.FC<Props> = ({
   open,
@@ -64,16 +59,11 @@ const ValueEditorDialog: React.FC<Props> = ({
   const handleDialogClose = () => setShowDialog(undefined);
   const [value, setValue] = useState(fieldContext.value);
 
-  // Template-related state
-  const [selectedTemplate, setSelectedTemplate] =
-    useState<ExpressionTemplate | null>(null);
-  const [showPlaceholderDialog, setShowPlaceholderDialog] = useState(false);
-
   // Fullscreen state
   const [isFullscreen, setIsFullscreen] = useState(false);
 
-  // Ref for RhaiCodeEditor to enable cursor insertion
-  const rhaiEditorRef = useRef<RhaiCodeEditorRef>(null);
+  // Ref for FlowExprCodeEditor to enable cursor insertion
+  const editorRef = useRef<FlowExprCodeEditorRef>(null);
 
   const [currentProject] = useCurrentProject();
 
@@ -136,30 +126,6 @@ const ValueEditorDialog: React.FC<Props> = ({
     handleDialogClose();
   };
 
-  // Template handlers
-  const handleTemplateSelect = useCallback((template: ExpressionTemplate) => {
-    setSelectedTemplate(template);
-    handleDialogClose();
-
-    // If template has placeholders, show placeholder dialog, otherwise replace directly
-    if (template.placeholders.length > 0) {
-      setShowPlaceholderDialog(true);
-    } else {
-      setValue(template.rhaiCode); // Templates replace entire content
-    }
-  }, []);
-
-  const handleTemplateInsert = useCallback((populatedCode: string) => {
-    setValue(populatedCode);
-    setShowPlaceholderDialog(false);
-    setSelectedTemplate(null);
-  }, []);
-
-  const handlePlaceholderDialogClose = useCallback(() => {
-    setShowPlaceholderDialog(false);
-    setSelectedTemplate(null);
-  }, []);
-
   const handleFullscreenToggle = useCallback(() => {
     setIsFullscreen((prev) => !prev);
   }, []);
@@ -183,13 +149,6 @@ const ValueEditorDialog: React.FC<Props> = ({
                   {fieldType ? `(${fieldType})` : ""}
                 </div>
                 <div className="flex flex-1 items-center gap-2">
-                  {/* <Button
-                    variant="outline"
-                    size="sm"
-                    onClick={() => handleDialogOpen("templates")}>
-                    <CodeIcon className="h-4 w-4" />
-                    {t("Templates")}
-                  </Button> */}
                   <Button
                     variant="outline"
                     size="sm"
@@ -255,16 +214,15 @@ const ValueEditorDialog: React.FC<Props> = ({
 
           <div
             className={`flex flex-col ${isFullscreen ? "h-[calc(100vh-52px)]" : "h-[600px]"}`}>
-            {/* Raw Rhai Editor - Always Visible */}
             <div className="relative flex-1 border-b">
-              <RhaiCodeEditor
-                ref={rhaiEditorRef}
+              <FlowExprCodeEditor
+                ref={editorRef}
                 className="h-full rounded-none bg-card/20 backdrop-blur-sm"
                 placeholder={t("Enter expression...")}
                 value={value}
                 onChange={setValue}
                 data-testid="value-editor-textarea"
-                aria-label={t("Raw Expression Editor")}
+                aria-label={t("Expression Editor")}
                 data-placeholder={t("Enter expression...")}
               />
               <Tooltip>
@@ -276,9 +234,7 @@ const ValueEditorDialog: React.FC<Props> = ({
                 <TooltipContent side="top" align="end">
                   <p className="text-sm">{t("Expression Editor Help")}</p>
                   <p className="mt-1 max-w-[200px] text-xs text-muted-foreground">
-                    {t(
-                      "Write Rhai expressions directly or use the visual builder below for assistance.",
-                    )}
+                    {t("Write FlowExpr expressions directly.")}
                   </p>
                 </TooltipContent>
               </Tooltip>
@@ -292,13 +248,6 @@ const ValueEditorDialog: React.FC<Props> = ({
           </div>
         </DialogContent>
       </Dialog>
-      {showDialog === "templates" && (
-        <TemplateLibraryDialog
-          open={showDialog === "templates"}
-          onClose={handleDialogClose}
-          onTemplateSelect={handleTemplateSelect}
-        />
-      )}
       {showDialog === "assets" && fieldContext && (
         <AssetsDialog
           onDialogClose={handleDialogClose}
@@ -311,12 +260,6 @@ const ValueEditorDialog: React.FC<Props> = ({
           onCmsItemValue={handleCmsItemValue}
         />
       )}
-      <TemplatePlaceholderDialog
-        open={showPlaceholderDialog}
-        template={selectedTemplate}
-        onClose={handlePlaceholderDialogClose}
-        onInsert={handleTemplateInsert}
-      />
     </>
   );
 };
