--- ui/src/features/Editor/components/ParamsDialog/components/ValueEditorDialog/templates/index.ts
@@ -1,8 +1,0 @@
-export { default as TemplateLibraryDialog } from "./TemplateLibraryDialog";
-export { default as TemplatePlaceholderDialog } from "./TemplatePlaceholderDialog";
-
-export * from "./templateData";
-export * from "./templateUtils";
-
-export type { ExpressionTemplate } from "./templateData";
-export type { PlaceholderValue, ProcessedTemplate } from "./templateUtils";
